# api_gateway/docs/swagger.yaml
basePath: /
definitions:
...
  dto.NewOrderItem:
    properties:
      menu_id:
        type: integer
      qty:
        type: integer
    required:
    - menu_id
    type: object
  dto.NewOrderRequest:
    properties:
      address:
        type: string
      items:
        items:
          $ref: '#/definitions/dto.NewOrderItem'
        type: array
    required:
    - address
    - items
    type: object
...
  dto.SwaggerResponese:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/orderpb.Orders'
        x-order: "1"
      message:
        type: string
        x-order: "0"
    type: object
...
  dto.SwaggerResponseDriverGetAllOrders:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/orderpb.Orders'
        x-order: "1"
      message:
        type: string
        x-order: "0"
    type: object
  dto.SwaggerResponseDriverGetCurrentOrder:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/orderpb.Order'
        x-order: "1"
      message:
        type: string
        x-order: "0"
    type: object
...
  dto.SwaggerResponseOrder:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/orderpb.Order'
        x-order: "1"
      message:
        type: string
        x-order: "0"
    type: object
...
  dto.SwaggerResponseUpdateOrder:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/orderpb.RequestUpdateData'
        x-order: "1"
      message:
        type: string
        x-order: "0"
    type: object
...
  dto.UpdateOrderStatus:
    properties:
      status:
        enum:
        - cancelled
        - done
        type: string
    required:
    - status
    type: object
...
  orderpb.Address:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    type: object
  orderpb.Driver:
    properties:
      id:
        type: integer
      name:
        type: string
      status:
        type: string
    type: object
  orderpb.Menu:
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
      subtotal:
        type: number
    type: object
  orderpb.Order:
    properties:
      driver:
        $ref: '#/definitions/orderpb.Driver'
      object_id:
        type: string
      order_detail:
        $ref: '#/definitions/orderpb.OrderDetail'
      payment:
        $ref: '#/definitions/orderpb.Payment'
      restaurant:
        $ref: '#/definitions/orderpb.Restaurant'
      user:
        $ref: '#/definitions/orderpb.User'
    type: object
  orderpb.OrderDetail:
    properties:
      created_at:
        type: string
      delivery_fee:
        type: number
      grand_total:
        type: number
      items_subtotal:
        type: number
      menu:
        items:
          $ref: '#/definitions/orderpb.Menu'
        type: array
      status:
        type: string
    type: object
  orderpb.Orders:
    properties:
      orders:
        items:
          $ref: '#/definitions/orderpb.Order'
        type: array
    type: object
  orderpb.Payment:
    properties:
      invoice_id:
        type: string
      invoice_url:
        type: string
      method:
        type: string
      status:
        type: string
      total:
        type: number
    type: object
  orderpb.RequestUpdateData:
    properties:
      order_id:
        type: string
      status:
        type: string
      user_id:
        type: integer
    type: object
  orderpb.Restaurant:
    properties:
      address:
        $ref: '#/definitions/orderpb.Address'
      admin_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      status:
        type: string
    type: object
  orderpb.User:
    properties:
      address:
        $ref: '#/definitions/orderpb.Address'
      email:
        type: string
      name:
        type: string
      user_id:
        type: integer
    type: object
  utils.ErrResponse:
    properties:
      detail:
        type: string
        x-order: "2"
      status:
        type: integer
        x-order: "0"
      type:
        type: string
        x-order: "1"
    type: object
host: jeki-x72poumuyq-et.a.run.app
info:
  contact:
    email: daniel.rahmanto@gmail.com
    name: Contact the developer
  description: Food delivery app built with microservices that integrates customer,
    driver, and restaurant.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Jeki
  version: "1.0"
paths:
...
  /driver/ongoing:
    get:
      description: Retrieve ongoing order related to the logged in driver. You will
        need an 'Authorization' cookie attached with this request.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SwaggerResponseDriverGetCurrentOrder'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Get current driver's order
      tags:
      - driver
  /driver/orders:
    get:
      description: Retrieve all orders related to the logged in driver. You will need
        an 'Authorization' cookie attached with this request.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SwaggerResponseDriverGetAllOrders'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Get all driver's orders
      tags:
      - driver
  /driver/orders/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve an order by it's ID. You will need an 'Authorization'
        cookie attached with this request.
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SwaggerResponseOrder'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Get order by ID
      tags:
      - driver
    put:
      consumes:
      - application/json
      description: Updates ongoing order status related to the logged in driver. You
        will need an 'Authorization' cookie attached with this request.
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Driver Update Order
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateOrderStatus'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SwaggerResponseUpdateOrder'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Update driver's order
      tags:
      - driver
...
  /merchant/ongoing:
    get:
      description: Retrieve ongoing orders for the logged in merchant admin. You will
        need an 'Authorization' cookie attached with this request.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SwaggerResponese'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Get ongoing merchant's order
      tags:
      - merchant
  /merchant/orders:
    get:
      description: Retrieve all orders for the logged in merchant admin. You will
        need an 'Authorization' cookie attached with this request.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SwaggerResponese'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Get all merchant's order
      tags:
      - merchant
  /merchant/orders/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve order details by it's order ID. You will need an 'Authorization'
        cookie attached with this request.
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SwaggerResponseOrder'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Get order by ID
      tags:
      - merchant
    put:
      consumes:
      - application/json
      description: Updates order status to the current logged in merchant admin. You
        will need an 'Authorization' cookie attached with this request.
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: string
      - description: Merchant Update Order
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateOrderStatus'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SwaggerResponseUpdateOrder'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Update merchant's order status
      tags:
      - merchant
...
  /users/ongoing:
    get:
      consumes:
      - application/json
      description: Get ongoing orders for the logged in user. You will need an 'Authorization'
        cookie attached with this request.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SwaggerResponese'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Get ongoing user orders
      tags:
      - customer
  /users/orders:
    get:
      consumes:
      - application/json
      description: Retrieves all orders for the logged-in user. You will need an 'Authorization'
        cookie attached with this request.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SwaggerResponese'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Get all user orders
      tags:
      - customer
    post:
      consumes:
      - application/json
      description: Create a new order for the logged in user. You will need an 'Authorization'
        cookie attached with this request.
      parameters:
      - description: Order details
        in: body
        name: orderRequest
        required: true
        schema:
          $ref: '#/definitions/dto.NewOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SwaggerResponseOrder'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Create new order
      tags:
      - customer
  /users/orders/{id}:
    get:
      consumes:
      - application/json
      description: Get order details by it's order ID. You will need an 'Authorization'
        cookie attached with this request.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SwaggerResponseOrder'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Get order by ID
      tags:
      - customer
...
swagger: "2.0"